requirements.txt
termcolor==1.1.0
certstream
tqdm
tld==0.7.9
levenshtein
websocket-client==0.48.0
PyYAML==5.1
collection
requests
bs4